reorg extended-styles
