+++ packages/material-ui/src/styles/extended-styles/create-with-extensions-rule.js
@@ -1,16 +1,16 @@
 import { StyleSheet } from 'react-native';
-import { addExtensions, excludeExtensions } from './extendedStyles';
+import * as Extensions from './extensions';
 
 function createWithExtensionsRule(styleSheet) {
   const styles = StyleSheet.create(
     Object.keys(styleSheet).reduce((stripped, styleName) => {
-      stripped[styleName] = excludeExtensions(styleSheet[styleName]);
+      stripped[styleName] = Extensions.exclude(styleSheet[styleName]);
       return stripped;
     }, {}),
   );
   Object.keys(styles).forEach(styleName => {
     // TODO meta for each extended style
-    addExtensions(styles[styleName], styleSheet[styleName]);
+    Extensions.add(styles[styleName], styleSheet[styleName]);
   });
   return { styles };
 }
