+++ packages/material-ui/src/styles/extended-styles/index.js
@@ -1,0 +1,4 @@
+import * as Animated from './animated';
+import render from './render-stylesheet';
+
+export { Animated, render };
