+++ packages/material-ui/src/styles/extended-styles/resolve-media-queries.js
