+++ packages/material-ui/src/styles/extended-styles/shorthand-properties.js
